refactor(useHotKey): migrate code to Typescript
