+++ src/components/NcAppNavigation/NcAppNavigation.vue
@@ -175,7 +175,7 @@ import Vue from 'vue'
 
 import { getTrapStack } from '../../utils/focusTrap.ts'
 import { logger } from '../../utils/logger.ts'
-import { useHotKey } from '../../composables/useHotKey/index.js'
+import { useHotKey } from '../../composables/useHotKey/index.ts'
 import { useIsMobile } from '../../composables/useIsMobile/index.js'
 
 import NcAppNavigationList from '../NcAppNavigationList/index.js'
