+++ src/globals.d.ts
@@ -19,6 +19,8 @@ declare module '*?raw' {
 
 declare global {
 	interface Window {
-		_nc_contacts_menu_hooks: { [id: string]: ContactsMenuAction },
+		_nc_contacts_menu_hooks: { [id: string]: ContactsMenuAction }
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		OCP: any
 	}
 }
